Add interop package
So far this only parses QMetrics for version 5, 6 and 7, and it has no
utility functions for summarising various stats. This will come soon.

Some basic tests have also been added, and so far I think it looks ok.
The main thing that is missing is to test the parsing of unbinned
quality data, but I haven't added any real data for that just yet. The
test data that is included is just the header and the first 100 records
from a couple of real runs.