Parse interop summaries
I ran into so many issues with parsing the interop files that I decided
to abandon that for now. Instead I do a syscall to the interop summary
binary and collect summary statistics that way. We're missing a few
pieces of information this way, but at least I can get a hold of other
stats that I failed to generate.