Add filtering of run table
This turned out pretty nicely. The filters belong in the table header,
and the filtering is applied as soon as any of the fields change. In
addition to this, I should also add some date filtering, and also
pagination.

Similar filtering should be added to the QC table, and here the charts
should be connected to this filtering as well so that whatever is
displayed in the table is also what is being visualised.