Add pagination for run table
This meant a rehaul of the results returned by the run service. It now
also returns metadata on the number of items available, including how
many pages there are available give the page size. If a page is out of
range, an error will be returned. If the page size is less than 1, all
results are returned.

The cleve service in the stacktstorm pack has to be updated accordingly.

Similar changes should be made to the QC table so that it can display a
paginated table, but the results in the chart still represents all the
data given the active filters.