Add pagination to the QC table
This also makes it so that the charts respond to the table filtering.
The table is paginated to only show a few samples at a time, but the
charts show the data for all samples that match the current filters that
have been defined.

The implementation also meant that the interaction with the database had
to be changed to include filtering and that kind of thing. This resulted
in the number of queries being reduced to a single one instead of making
a query for each sample that has the ready state. Furthermore, a sample
that is ready does not necessarily have QC data yet, and by querying the
QC collection directly this is no longer a problem.