Database service and CLI for samplesheets
The current implementation allows for only one samplesheet per run, but
they can be updated. If a samplesheet is being added to a run with an
already existing samplesheet, the modification time of the file is taken
into account. If the modification time of the file is newer than what
currently exists in the database, the samplesheet information is
updated.