Implement charts on the run page
At the moment this is kind of rudimentary, but it's a decent start. Some
things that should be modified is to have a separate template for the
charts that can be controlled with variables passed in, and the
variables that should be possible to control should be populated based
on what is available in the database.