Refactor database access
Instead of having a service associated with each collection, the methods
for interacting with the database are defined directly on the DB struct.
The idea is to have the http handlers accept interfaces limited to the
methods that they will actually use. This will make testing easier and
cleaner than it is now.