refactor: better database interfaces (#18)
* Refactor database access

Instead of having a service associated with each collection, the methods
for interacting with the database are defined directly on the DB struct.
The idea is to have the http handlers accept interfaces limited to the
methods that they will actually use. This will make testing easier and
cleaner than it is now.

* Have http handlers accept interfaces

These might have to be modified for clarity further down the road, but
it's a start.

* Create new mocks and update tests

* Use absolute paths to test data

Ideally, test data should be deployed somewhere. Then either the tests
are run in a specific environment, or the data is downloaded if these
things should be tested.

* Add some more documentation

In general, I need to document the exported package members much better
than I'm doing at the moment. This will have to happen as I go, and I
will not attempt to do that in this PR.