Add handler for adding sample sheets without run ID
This will act on a separate endpoint from the sample sheets associated
with a run. It is still possible to supply a run ID at this stage, but I
would need to write some more tests for this. Meanwhile I would
recommend to use the run-specific endpoint for adding associting a
sample sheet with a run.