Only update run ID for sample sheet with existing UUID
If a sample sheet with the same UUID as the input sample sheet is found,
then only the run ID of the sample sheet gets updated. Otherwise the
whole sample sheet is updated, assuming that it is more recent than the
sample sheet in the database.

This approach is problematic though since it basically means that sample
sheets can never be updated. I will work on a different solution...