First attempt at a sample metadata section
Here I use a constant to define what section to look for, and then the
Validate method (previously IsValid) checks if it's present, and then
checks for the required columns. The required columns are very much a
work in progress, and will likely change multiple times in the near
future.

Right the section and the required columns are hard coded, which has
both pros and cons. The con is that it's pretty static in what it will
be looking for, and changing this requires a rebuild. At the same time I
would say that this is a pro since it would create a tighter connection
between the version of the program and what it actually requires.

Something I haven't done so far is to document things like this. This
should really be done, and I suppose this type of documentation could go
into the web interface. That would make things coherent, and you would
be able to find everything in one place.