Add new samplesheet representation to runs
Next would be to make sure that the representation is consitent when
querying a single run. The best way of doing this would be to simply
call `Runs` from `Run`, but this creates a small issue with the
filtering. The run_id is currently matched with a regex, and this would
break if there are runs whose IDs are prefixes of run IDs of other runs.
I would then have to introduce two separate fields for the filtering:
one that is a regex match and one that should be a strict match.