feat: manage sample sheets with UUIDs (#28)
* Get UUID from sample sheet RunDescription

* Change names of current sample sheet handlers

This to reflect that they act on sample sheets associated with
sequencing runs.

* Add skeletons for sample sheet endpoints

* Use functional options pattern for sample sheets

This makes it possible to create and query sample sheets either on run
ID or on UUID.

* Add tests for sample sheet options

* Fix bug with single-column sample sheet settings

* Have UUID be a field of SampleSheet

* Index on UUID and prioritise run ID in lookup

* Remove unnecessary prints

* Add sample sheet handler test

More tests are needed, but this is a start

* New function for writing temp files for tests

* Update sample sheet database index

This allows for missing values both when it comes to run IDs and UUIDs,
but if they are not missing, then duplicates are not allowed.

* Represent sample sheet run ID as pointer to string

This in order to allow for it to be represented as `nil`/`null` in cases
where there is a UUID present, but no run ID.

* Add handler for adding sample sheets without run ID

This will act on a separate endpoint from the sample sheets associated
with a run. It is still possible to supply a run ID at this stage, but I
would need to write some more tests for this. Meanwhile I would
recommend to use the run-specific endpoint for adding associting a
sample sheet with a run.

* Set run ID for sample sheet if provided

* Only update run ID for sample sheet with existing UUID

If a sample sheet with the same UUID as the input sample sheet is found,
then only the run ID of the sample sheet gets updated. Otherwise the
whole sample sheet is updated, assuming that it is more recent than the
sample sheet in the database.

This approach is problematic though since it basically means that sample
sheets can never be updated. I will work on a different solution...

* Add method for merging sample sheets

* New representation of paths and mod times

This now allows for a sample sheet to be composed of more than one
sample sheet. This is something that would only occur if two sample
sheets have been merged.

Next in this implementation would be to update the functions and structs
that interact with the database.

* Merge sample sheets if existing is found in database

Right now there are two representations of the sample sheet co-existing
in the database, but the model only represents the updated version
meaning that the file metadata will be missing for the old
representation until all sample sheets in the database have been
updated.

* Don't add duplicates to sample sheet files

* Omit empty sample sheets from runs

These are only populated by a database lookup, and will thus never
actually be populated in the run document, so it doesn't really make
sense to have it in there.

* Sort sample sheet files by modification time

This is partly to make the tests stable, but it also brings some
consistency since the most recent sample sheet will always be the last
element of the slice.

* Check UUID before getting existing sample sheets

* Test fail for empty sample sheet

* Correctly merge against empty sample sheets

* Merge on run ID if UUID is missing

* Add new samplesheet representation to runs

Next would be to make sure that the representation is consitent when
querying a single run. The best way of doing this would be to simply
call `Runs` from `Run`, but this creates a small issue with the
filtering. The run_id is currently matched with a regex, and this would
break if there are runs whose IDs are prefixes of run IDs of other runs.
I would then have to introduce two separate fields for the filtering:
one that is a regex match and one that should be a strict match.

* Utilise Runs when getting a single Run

This eliminates any discrepancies between single runs and multiple runs,
and makes it much easier and less error prone to update this query.

* Translate the old sample sheet to the new type