Drop platform collection and get them from the runs
This also adds a more feature rich platform handling in order to be able
to match the platforms across different versions of the run schema. I'm
not a huge fan of this current solution since it adds a bit of
complexity, but it works so I will just roll with it for now.