Set the run schema version when adding runs
When removing the schema version from the run, I didn't have the
adding of runs in mind. The schema version for new runs needs to be
added, but it doesn't need to be in the run struct. I now created an
auxiliary struct in the database function where I basically hard code
the version to the most recent one. This way no other part of the
application has to care about this.