feat: add interop package (#60)
* Add interop package

* ci: skip tests if files are not available

* Rename test data directories for consistency

* Add run ID to run info

* Identify platform and flowcell from IDs

* Split run info parsing into read and parse functions

* Represent dates as time.Time

* Make consumable public

* Add struct tags to run info and parameters

* Parse into bool instead of interopBool

* Make identify functions public

* Make FlowcellInfo public

* Use ReadInfo for both run info and paramters

* Drop read information from run parameters

This information is available in the run info, and with more details. No
point in keeping multiple copies of it.

* Update run struct and add support for old version

* go mod tidy

* Don't have formatters edit files

It creates a lot of annoying issues with the file being edited by the
formatter before I save, saying that the file has been modified on disk.
Better to just run the formatter myself before saving.

* Fix flowcell identification bug

There was an issue with some NextSeq IDs being identified as MiSeq
flowcells due to it matching both patterns. Due to maps being unordered,
the error was intermittent. Putting the patterns in a slice instead, and
moving the more ambiguous patterns to the end should resolve this issue.

* Use interop for parsing run info/parameters

This makes it possible to drop the run info/parameters from the cleve
package. Should be completely backwards compatible.

* Add platform to database when run is loaded

* Proper date sorting for schema version 1 and 2

* Drop platform collection and get them from the runs

This also adds a more feature rich platform handling in order to be able
to match the platforms across different versions of the run schema. I'm
not a huge fan of this current solution since it adds a bit of
complexity, but it works so I will just roll with it for now.

* Document the platform implementation

* Sort platforms by name

* Add run number to run info

This also fixes an issue with the unmarshalling of the run parameters
for schema version 1.

* Remove schema version from run struct

This is only relevant in context of the database. Downstream all the
data will have the same structure, and the schema version is irrelevant.

* Add ready marker to condensed platforms

* Fix platform JSON representation

* Remove mention of platform POST from API docs

* Set the consumable mode to string

There was some discrepancy in how this was represented, and I don't
really know what it's used for, so it's safest to have it as a string
for now. I might even remove it later on if we don't find a use for it.

* Fix some qc issues

Can't really test these right now, and they will change in the future.
At least the QC table displays correctly now.

* Set the run schema version when adding runs

When removing the schema version from the run, I didn't have the
adding of runs in mind. The schema version for new runs needs to be
added, but it doesn't need to be in the run struct. I now created an
auxiliary struct in the database function where I basically hard code
the version to the most recent one. This way no other part of the
application has to care about this.

* Remove stray logging

* Update QC platform filter to new structure