Schema versioning for the QC documents
This includes detecting this in the handler for single runs in order to
be able to notify the user of this while still being able to see the
page. For the QC table this will have to be taken care of in the
template. Not the most elegant solution, but it will work for now.