Only plot the current QC table view
This gets around an issue where the query would fail due to too much
data in a single document, which is a whole other can of worms that I
will have to open at some point. Ideally I should be able to plot
virtually any number of runs. The real problem is still there underneath
it all, but now it is much less likely to be triggered.