feat: more flexible interop handling (#62)
* Simplify some checks

* CLI flag for setting run state when adding run

* Fix issue with page size filter

The page size wasn't passed when clicking the pagination buttons, and
with the defaults you would go out of range in certain situations.

* Add command for initialising the database

This will add collections and set the correct indexes. The
CreateCollection function should error if the collection already exists,
but this does not happen for me. Everything seems to work though, but
it's a bit annoying that I can't get the right messages to display.

* Use the interop package for all the QC metrics

There are quite a few implementations that are missing at the moment,
but at least there is no longer any dependency on the Illumina interop
binaries (I think), so these will be removed in the next commit.

* Remove old interop implementation

Everything compiled fine without them

* Skip test if files no found

* Use struct instead of interface for QMetricRecord

* Add bin definitions for QMetrics V4

* Add percent aligned and fix some issues

* Functions for Q30

* Error rate functions

* Q30 -> PercentQ30

* Function for percent occupied

* QC run summary function

* QC tile summary function

* interface{} -> any

* Remove unused functions

* BSON and JSON naming

* Add percent passing filter to run summary

* Add index metrics to summary

* Add cluster density to lane summary

* Add read summary metrics

* Add cluster density to run summary

* Make lane summary a slice instead of map

* Make samples have the original order

This order might still be different from the one in the sample sheet,
but this is still better than the random ordering I had before.

* Have read stats be read->lane like everything else

* Better representation of missing values

* Fix bug in calculation of aligned per lane

* Schema versioning for the QC documents

This includes detecting this in the handler for single runs in order to
be able to notify the user of this while still being able to see the
page. For the QC table this will have to be taken care of in the
template. Not the most elegant solution, but it will work for now.

* Set percent occupied to NaN if extended tile metrics are missing

* Update chart values

* Update the UI with all the new data

* Handle NaNs in JSON marshalling

This is an ugly solution, but I can't think of anything better at the
moment. I'll just leave it like this for now. Everything seems to be
working as it should at least.

* Sort before pagination

* CLI command for updating QC data for a run

* Allow for globbing when looking for interop files

* Don't show indexing table if there is no info

* Issue with weird sorting of many runs with old QC

* Make sure path is absolute and directory

* Only outdated QC message if there actually is QC data for the run

* Only plot the current QC table view

This gets around an issue where the query would fail due to too much
data in a single document, which is a whole other can of worms that I
will have to open at some point. Ideally I should be able to plot
virtually any number of runs. The real problem is still there underneath
it all, but now it is much less likely to be triggered.

* Persist chart settings on navigation

* Tailwind update