Proper status and result when filter returns nothing
This includes missing platforms, which results in an early return. I
then set the page to 1 and the total number of pages to 1. If a page
number higher than 1 is set in the filter, this is essentially ignored,
and the URL will be a bit wonky. This should be fixed.