fix: better status codes for various server side errors (#84)
* Better error when accessing a non-existent run

* Proper status and result when filter returns nothing

This includes missing platforms, which results in an early return. I
then set the page to 1 and the total number of pages to 1. If a page
number higher than 1 is set in the filter, this is essentially ignored,
and the URL will be a bit wonky. This should be fixed.

* Indicate when the run table is empty

* Correct status for API request with filter that returns nothing

* Better debouncing of run ID search

* Proper status for QC filter that yields no results

* Proper page numbers for empty QC table

* Better debouncing for run ID search in QC table

* Display message for empty QC table

* Remove stray print

* Don't try to plot if there is not data to plot

* Populate the table select headers with custom values

If the value for platform or state are not known beforehand, still allow
them to be populated in the header to make it more cleare what the table
results are actually representing.

* Move platform name resolution to filter creation

* Return 404 if platform doesn't exist

I guess this error could happen in other cases as well. All these things
should probably be cleaned up a bit so that it's clear what the error
represents also in the code, not only in the response you get from the
server.