feat: add handling of gene panels (#88)
* Add basic panel representation and parsing

* Add date, categories and archive status to panels

* Split out panel version into its own struct

* Use `cleve_version` instead of `version` to avoid confusion

The version is also included in the response headers, so I could
possibly use this instead to keep things a bit more clean. I would still
have to handle this in the handlers though, so it wouldn't make too much
of a difference.

* Include header data in error pages

* Add a panel filter

* Add database logic for gene panels

* Add routes and templates for gene panels

* Style HTTP 500 page to be consistent with other pages

* Ability to (un)archive a panel from the command line

* Add archive status to panel info template

* Fix error when trying to access non-existing panel without version

* If archived is true, all panels should be shown

* Add CLI command for listing panels

* Italicise gene symbols

* Add time when panel was archived

* Present dates in local time zone

* Return empty slice if no panels are found

* More consistent naming of filter fields

* Add API endpoints for getting gene panels

* Forgot to add handlers for API endpoints

* No error if there are no panels in the database

* Use custom time for panel date

* Add panel validation

* Add API endpoint for creating a gene panel

Currently JSON is the only accepted input format. This might change in
the future, but maybe not. I still have to figure out whether it's the
responsibility of Cleve to check that gene IDs are valid, or if that
should be done by the API consumer.

* Add CLI command for deleting panels

* Ditch the custom time

Custom times cause more headaches than not. Now the time is stored as
time.Time both in the database and in the struct. Parsing issues are
then left to the http handler and the text parser.

* Make version variables package private

* Add version struct with tests

This will initially be used for the panels, but could potentially be
used in more places.

* Zero version string should be ""

* Implement the JSON Marshaler interface for Version

* Add the Version struct to the panel handling

* Add endpoint for archiving a panel

* Don't allow adding panels in certain cases

Panels should not be allowed to be added to the database if:

- The creation date is in the future
- The panel has been archived
- There is a newer panel in the database already

* Add methods for comparing versions

* Don't add panels with a version older than the most recent

* Use int for HGNC ID

* Check that all genes have an HGNC ID in the validation