Keep ordering of scatter plot series consistent
I did this by making the scatter data into a generic type constrained on
the group field that has to implement the cmp.Ordered interface. This is
not a great solution, at least in terms of ergonomics, but I couldn't
really find a better way of doing it. I'll keep it for now since it
seems to be working fine.