feat: simplify run update endpoint (#109)
* Single endpoint for updating runs

* Update analyses if they are subdirectories of the run

If an update of the run path is requested and the run has associated
analyses, we go through all the analyses and check if the original
analysis path is a subdirectory of the original run directory. If that
is the case, then the path is updated. If not it is left alone.

This makes quite a few assumptions that aren't necessarily true, and the
handling of this should ideally be rewritten in a more solid manner, but
this will have to do for now. One of the caveats is that if the updating
of an analysis path fails, all preceding operations have already taken
effect while any coming operations are ignored.

* Update API documentation

This deprecates the two previous endpoints for updating the run state
and the run path.