Better state update logic
If the state was given on the command line for the update, then use this
state regardless of what the actual state of the run is. If the path has
been updated, then do a forceful check of the state of the run and use
this for the update. In all other cases detect the current state of the
run and update it if it differs from the last known state.