fix: better run state update logic (#119)
* Setup default slog instance

* Add force flag to state detection

* Add more state tests

* Add some more logging to the run watcher

* Check that run ids match when updating path

* Better state update logic

If the state was given on the command line for the update, then use this
state regardless of what the actual state of the run is. If the path has
been updated, then do a forceful check of the state of the run and use
this for the update. In all other cases detect the current state of the
run and update it if it differs from the last known state.

* Simpler state update logic

The ignored states are checked in the Run.State method, so no need to
duplicate it here.

* log -> log/slog

* Allow for empty update payload

* Check that run IDs match before updating

* Force state update if path is updated

* Update API documentation

* Simplification suggest by linter