fix: better watcher implementation (#122)
* Rename RunState to State

This is more general and can be used for other things, so it makes sense
to call it something a bit more general.

* Simplify StateHistory.LastState to only return the state

The time was never used anyways, and the method name suggests that only
the state would be returned.

* Rename run_watcher.go -> watcher.go

* Fix poll interval parameter default and precedence

* Don't update state in the watcher, emit on channel instead

* Add id field to WatcherEvent struct

* Don't emit events if state has not changed

* Handle run watcher events

* Use specific field for state change

This will enable the addition of other fields in the struct, and
different updates can be triggered even if the state hasn't necessarily
changed.

* Refactor watcher methods

* Remove unused method from interface

* More test cases and better tests

* Default copy complete and run completion files

These used to return empty strings which would result in misinterpreting
the state of runs. Mostly an issue in testing.

* Better understandable if statement

* Proper run pagination in test

* Add test for moved run