feat: more general handling of analyses (#123)
* Dragen analysis implementation

This is far from done, but it's a start. The initial parsing should
actually generate both a run-level analysis with the stats from
BCLConvert, as well as per-sample analyses that each contain the fastq
files associated with those samples.

* fix: update state handling in templates (#124)

* Make DragenManifest a struct and add FindFiles method

* Create analyses collection

* Marshalling and unmarshalling for AnalysisLevel

* Update gin to v1.10.1

This is mostly to get the UnmarshalParam interface that was introduced
in v1.10.0 that will make it possible to unmarshal query params into
custom types.

* Add UnmarshalParam methods to state and level

* Add an analysis filter

* Add AnalysisResult struct

* Update mongo analyses function

* Update mongo analysis function

* Add SetAnalysesIndex function

* Update handlers for run-associated analyses

* Better default handling for pagination filter

* Make the default state and level invalid

* Don't use pointers for state and level

* Use blank identifier for invalid state and level

* Adding analyses should happen under analyses, not runs

Far from all analyses will be associated with runs, and therefore it
makes more sense to add analyses in a more general way.

* Get analyses from multiple endpoints

* Make files an empty array if nil

* Return a slice of Dragen analyses

One run level analysis with stats and one sample level analysis per
sample.

* Delete commented function

* Fix analysis update endpoints

* Set database created and updated time for analyses

* Remove analysis from run struct

* Ignore analyses in the run documents

* Bump run schema version

* Update API docs

* Remove debug log

* Clarify how the files array should look

* Fix parameters when filtering analyses without a level

* That's random

* Add tests for dragen analysis

* Fix bugs in loading dragen analysis

* Fix test issues and more complete tests