Move analysis level to output files
Files have been renamed OutputFiles in the Analysis struct, and the
Level and ParentId fields have been moved to the files. This allows for
having a single analysis represent files that are associated with
different levels (run, case, sample). Also, the field Runs was added to
the Analysis struct, and this should represent what runs the analysis is
associated with. For Dragen analyses, this should be a single run at all
times, while for downstream analyses, this could represent more than one
run. For example, if a sample has been resequenced due to not having
enough data for the analysis, it will be associated with at least two
runs (one with too little data, and one with the extra sequencing).