feat: even more general handling of analyses (#132)
* Move analysis level to output files

Files have been renamed OutputFiles in the Analysis struct, and the
Level and ParentId fields have been moved to the files. This allows for
having a single analysis represent files that are associated with
different levels (run, case, sample). Also, the field Runs was added to
the Analysis struct, and this should represent what runs the analysis is
associated with. For Dragen analyses, this should be a single run at all
times, while for downstream analyses, this could represent more than one
run. For example, if a sample has been resequenced due to not having
enough data for the analysis, it will be associated with at least two
runs (one with too little data, and one with the extra sequencing).

* Use AnalysisFileFilter for getting analysis files

* Update docstring

* Add FindFile method to DragenManifest

* Use manifest to get demux stats

* Correct json and bson name

* Ensure analysis path is absolute

* Add file name to AnalysisFileFilter

* Add run ID to analysis filter

* Make sure files are empty arrays if missing

* Add optional run ID constraint when fetching single analysis

This will make it possible to restrict the analyses that will be
returned when checking for analyses that are associated with a certain
run from the `/api/runs/:runId/analyses/:analysisId` endpoint.

* Add endpoint for getting single analysis

* Get and set index on analyses collection

* Ditch parent ID in analyses endpoints

This is no longer a thing

* Clear distinction between analysis input and output files

* Add tags to AnalysisFileFilter

* Move AnalysisFileFilter to filters

* Add form tags

* Unmarshal empty string level without error

* Check files for completeness when updating

* Check output files for completeness when adding analysis

* Update marshalling tests for AnalysisLevel

* Add IsValid method to AnalysisFileFilter

* More tests

* IsValid/IsComplete -> Validate

* Use value receiver for MarshalJSON

Fell for this. Again.

* The bson package uses omitempty, not omitzero

* Check that input files definitions are valid

* Update API docs