feat: API endpoints for getting analysis files (#137)
* Add UnmarshalParam method for AnalysisFileType

* Add endpoint for getting output files from analyses

* Allow analysis id to be missing

In cases where the id cannot be missing, this will have to be checked
for explicitly.

* Return []AnalysisFile instead of []string

* Check analysis file validity

This test will fail right now since these paths are absolute (which is
very much intended). I need to fix this by having some kind of
indication whether or not the analysis file is directly associated with
an analysis.

* Add flag indicating association with an analysis

* Add endpoint for getting the files for a specific analysis

This isn't entirely necessary since this is also possible to do through
`/api/analyses?analysis_id={analysis_id}`. Might still be useful and
won't create too much more to maintain.

* Return ErrNoDocuments if analysis ID and it's not found

* Respond with 404 if analysis is not found

* Don't require the analysis level to be set

* Update API documentation